modified:   client.py 	modified:   data2.py 	deleted:    layout.py 	modified:   main.py 	modified:   static/styles.css
